(chore): cluster shows not ready when node unschedulable

(chore): set ready false when all nodes are not ready


(chore): e2e test node not ready


(chore): set not ready only when all nodes are not ready


(chore): set PayloadSchedulable condition


Automatic generation of CRD API Docs

+++ internal/controller/cluster/status_test.go
@@ -155,9 +155,9 @@ var _ = Describe("Cluster status", Ordered, func() {
 			g.Expect(validCluster.Status.GetConditionByType(greenhousev1alpha1.AllNodesReady).Status).To(Equal(metav1.ConditionFalse))
 			g.Expect(validCluster.Status.GetConditionByType(greenhousev1alpha1.AllNodesReady).Message).To(ContainSubstring("test-node not ready, test-node-3 not ready"))
 			// Validate the counts of total and ready nodes.
-			g.Expect(validCluster.Status.Nodes).ToNot((BeNil()))
-			g.Expect(validCluster.Status.Nodes.Total).To(Equal(int32(3)))
-			g.Expect(validCluster.Status.Nodes.Ready).To(Equal(int32(1)))
+			g.Expect(validCluster.Status.Nodes).ToNot(BeNil())
+			g.Expect(validCluster.Status.Nodes.Total).To(Equal(3))
+			g.Expect(validCluster.Status.Nodes.Ready).To(Equal(1))
 			g.Expect(validCluster.Status.Nodes.NotReady).To(HaveLen(2))
 		}).Should(Succeed())
 
@@ -217,9 +217,9 @@ var _ = Describe("Cluster status", Ordered, func() {
 			g.Expect(validCluster.Status.GetConditionByType(greenhousev1alpha1.AllNodesReady).Status).To(Equal(metav1.ConditionTrue), "The AllNodesReady condition should be true")
 			g.Expect(validCluster.Status.GetConditionByType(greenhousev1alpha1.AllNodesReady).Message).To(BeEmpty())
 			// Validate the counts of total and ready nodes.
-			g.Expect(validCluster.Status.Nodes).ToNot((BeNil()))
-			g.Expect(validCluster.Status.Nodes.Total).To(Equal(int32(3)))
-			g.Expect(validCluster.Status.Nodes.Ready).To(Equal(int32(3)))
+			g.Expect(validCluster.Status.Nodes).ToNot(BeNil())
+			g.Expect(validCluster.Status.Nodes.Total).To(Equal(3))
+			g.Expect(validCluster.Status.Nodes.Ready).To(Equal(3))
 			g.Expect(validCluster.Status.Nodes.NotReady).To(BeEmpty())
 		}).Should(Succeed())
